app/requirements.txt
tensorflow-cpu==2.8.0
pillow
protobuf==3.20.*